test(mister): isolate backup mount smoke test
